domain - cluster vulnerability and cluster sbom factory

+++ src/TrivyOperator.Dashboard/Infrastructure/Trivy/Schema/ReportSchemas/Sboms/SummaryCr.cs
@@ -5,8 +5,8 @@ namespace TrivyOperator.Dashboard.Infrastructure.Trivy.Schema.ReportSchemas.Sbom
 public class SummaryCr
 {
     [JsonPropertyName("componentsCount")]
-    public long ComponentsCount { get; init; } = 0;
+    public int ComponentsCount { get; init; } = 0;
 
     [JsonPropertyName("dependenciesCount")]
-    public long DependenciesCount { get; init; } = 0;
+    public int DependenciesCount { get; init; } = 0;
 }
